+++ content/1.Domain/Company.Project.Domain.Interfaces/Generics/Base/IBaseService.cs
@@ -36,7 +36,7 @@
         /// </summary>
         /// <param name="id">The identifier.</param>
         /// <returns></returns>
-        Task<TEntity?> Read(ulong id);
+        Task<TEntity?> Read(uint id);
 
         /// <summary>
         /// Reads with paged results.
@@ -60,6 +60,6 @@
         /// </summary>
         /// <param name="id">The identifier.</param>
         /// <returns></returns>
-        Task<bool> Delete(ulong id);
+        Task<bool> Delete(uint id);
     }
 }
